Add ResolverFactory
observation 1:
the `StorageCredentialCache` in `PolarisEntityManager` was only ever used
by `PolarisEntityManager.getCredentialCache()` which was only ever called
by `FileIOUtil.refreshAccessConfig`.
so instead we provide `DefaultFileIOFactory` and `IcebergCatalog` with the
`StorageCredentialCache` directly.

observation 2:
`PolarisEntityManager.prepareResolver` was the only place that used the
`EntityCache`.
By introducing an application-scoped `ResolverFactory` the creation of
an `EntityCache` is decided only in `QuarkusProducers`, which has become
the only place where `MetaStoreManagerFactory.getOrCreateEntityCache` gets called.